Split test/runners/javascript_spec.js
