+++ test/runners/javascript/examples.js
@@ -1,0 +1,7 @@
+"use strict";
+
+const runner = require('../../runner');
+
+describe('examples', function() {
+  runner.assertCodeExamples('javascript');
+});
